refactor: Wave 5 mechanical — Java breadcrumb strip + io_loop verify + api-redesign reconcile + MSRV + semver-checks CI (refs #500)
Five mechanical items batched into one PR + one 9.0.1 patch bump.

A. **Java RE breadcrumbs.** Stripped 76 per-line `// Source:
   PojoMessageUtils...` / `// Java: ...` / `(FPSSClient.foo())`
   references across `crates/`, `tools/`, `ffi/`. ADR-001
   (`docs/architecture/ADR-001-java-terminal-parity.md`) is the
   institutional anchor; the per-line refs were dead scaffolding from
   the original RE pass. Module-header parity prose collapses to a
   single `See ADR-001 (...)` line in `fpss/mod.rs`,
   `fpss/connection.rs`, `fpss/protocol/mod.rs`,
   `fpss/protocol/wire.rs`, `mdds/decode/error.rs`, and
   `config/mdds.rs`. Final scrub returns zero hits — well under the
   audit's "≤ 5" budget. Three contract codec tests rename
   `java_parity_*` → `wire_parity_*`.

B. **`io_loop.rs` (audit §2.9).** Three concerns: login handshake,
   main blocking-read + Disruptor-publish + command-drain loop, ping
   heartbeat scheduler. Split into
   `io_loop/{mod.rs, login.rs, ping.rs}`. login + ping each lift with
   their own state and tests; the main loop body stays in mod.rs
   because it owns the session state machine (frame buffer,
   `FrameReadState`, delta state, contract cache, reconnect attempt
   counter, Disruptor producer) which doesn't decompose without
   leaking state across module boundaries. `pinning.rs` has two
   callsites (`fpss/connection.rs` + `flatfiles/mdds_spki.rs` reuses
   the `FPSS_SPKI_SHA256` digest), so it stays separate — folding it
   into `io_loop` would force every flatfiles consumer to import the
   FPSS I/O module purely for the SPKI constant.

C. **`docs/public-api-redesign.md`.** Fully superseded by v9.0.0:
   the two-layer exact + ergonomic surface, typed scalars (`Date`,
   `Year`, `Right`, `Expiration`, `Strike`, `TimeOfDay`, `Interval`),
   `IntoOptionSpec`, `for_contracts(...)`, and `StrikeSelector` /
   `ExpirationSelector` wildcards all shipped on the 9.0.x line. The
   doc's proposed cadence ("v6.x additive façade ... v8.0 remove
   compat") is moot. Deleted; new `docs/architecture/README.md`
   indexes the ADR set and records that planning prose for shipped
   surfaces is intentionally not preserved.

E. **MSRV.** True floor is 1.88 (not the prior `1.85` declaration —
   transitive deps `time 0.3.47`, `time-core 0.1.8`, `time-macros
   0.2.27`, the `icu_*` 2.2.0 family already require `>= 1.86` /
   `>= 1.88`). Bumped `[workspace.package].rust-version = "1.88"`,
   the four excluded packages (`tools/server`, `tools/mcp`,
   `sdks/python`, `sdks/typescript`) declare `rust-version = "1.88"`
   directly. CI `Lint` job grows a `toolchain` matrix axis pinned to
   `stable + 1.88` on Linux. `README.md` adds a `Requirements`
   section. While the floor is touched, `tdbe::codec::fie` swaps
   `len % 2 == 0` for `len.is_multiple_of(2)`, which stable
   clippy 1.94 flags as `manual_is_multiple_of`.

F. **`cargo-semver-checks` CI gate.** New `Semver check` job runs
   `obi1kenobi/cargo-semver-checks-action@v2` against the `v9.0.0`
   tag on every PR. Verified locally — both `tdbe` (196 checks) and
   `thetadatadx` (222 checks) pass with zero breakage and no
   false-positives, so no `#[doc(hidden)]` suppressions needed.
   `CONTRIBUTING.md` documents the public-API stability policy and
   the local invocation.

Refs #500.
