+++ .github/workflows/ci.yml
@@ -43,18 +43,29 @@ jobs:
           arguments: --all-features
 
   check:
-    name: Lint (${{ matrix.os }})
+    name: Lint (${{ matrix.os }}, ${{ matrix.toolchain }})
     strategy:
       fail-fast: false
       matrix:
         os: [ubuntu-latest, macos-latest, windows-latest]
+        toolchain: [stable]
+        include:
+          # MSRV floor declared in `[workspace.package].rust-version`.
+          # Pin the Linux runner to the floor toolchain so a dependency
+          # bump that pulls in a newer rustc requirement fails CI before
+          # it reaches release.
+          - os: ubuntu-latest
+            toolchain: "1.88"
     runs-on: ${{ matrix.os }}
     timeout-minutes: 30
     steps:
       - uses: actions/checkout@v6
-      - uses: ./.github/actions/setup-rust-deps
+      - uses: dtolnay/rust-toolchain@master
         with:
+          toolchain: ${{ matrix.toolchain }}
           components: clippy
+      - uses: Swatinem/rust-cache@v2
+      - uses: ./.github/actions/install-protoc
       - run: cargo clippy --workspace --locked -- -D warnings
 
   test:
@@ -66,6 +77,18 @@ jobs:
       - uses: actions/checkout@v6
       - uses: ./.github/actions/setup-rust-deps
       - run: cargo test --workspace --locked
+
+  semver:
+    name: Semver check
+    runs-on: ubuntu-latest
+    needs: check
+    timeout-minutes: 30
+    steps:
+      - uses: actions/checkout@v6
+      - uses: dtolnay/rust-toolchain@stable
+      - uses: obi1kenobi/cargo-semver-checks-action@v2
+        with:
+          baseline-rev: v9.0.0
 
   surfaces:
     name: Extended Surfaces
