+++ sdks/typescript/npm/darwin-arm64/package.json
@@ -1,6 +1,6 @@
 {
   "name": "thetadatadx-darwin-arm64",
-  "version": "9.0.0",
+  "version": "9.0.1",
   "os": [
     "darwin"
   ],
