+++ sdks/typescript/npm/linux-x64-gnu/package.json
@@ -1,6 +1,6 @@
 {
   "name": "thetadatadx-linux-x64-gnu",
-  "version": "9.0.0",
+  "version": "9.0.1",
   "os": [
     "linux"
   ],
