+++ sdks/typescript/npm/win32-x64-msvc/package.json
@@ -1,6 +1,6 @@
 {
   "name": "thetadatadx-win32-x64-msvc",
-  "version": "9.0.0",
+  "version": "9.0.1",
   "os": [
     "win32"
   ],
