+++ sdks/typescript/package.json
@@ -1,6 +1,6 @@
 {
   "name": "thetadatadx",
-  "version": "9.0.0",
+  "version": "9.0.1",
   "description": "Native ThetaData SDK for Node.js — powered by Rust via napi-rs",
   "license": "Apache-2.0",
   "repository": {
@@ -30,9 +30,9 @@
     "@napi-rs/cli": "^3.6.2"
   },
   "optionalDependencies": {
-    "thetadatadx-linux-x64-gnu": "9.0.0",
-    "thetadatadx-darwin-arm64": "9.0.0",
-    "thetadatadx-win32-x64-msvc": "9.0.0"
+    "thetadatadx-linux-x64-gnu": "9.0.1",
+    "thetadatadx-darwin-arm64": "9.0.1",
+    "thetadatadx-win32-x64-msvc": "9.0.1"
   },
   "engines": {
     "node": ">= 20"
