+++ tools/cli/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "thetadatadx-cli"
-version = "9.0.0"
+version = "9.0.1"
 edition.workspace = true
 rust-version.workspace = true
 authors.workspace = true
