+++ tools/mcp/Cargo.toml
@@ -1,7 +1,8 @@
 [package]
 name = "thetadatadx-mcp"
-version = "9.0.0"
+version = "9.0.1"
 edition = "2021"
+rust-version = "1.88"
 description = "MCP server for ThetaDataDx — gives LLMs instant access to ThetaData market data"
 license = "Apache-2.0"
 repository = "https://github.com/userFRM/ThetaDataDx"
